+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## **unreleased** - unreleased
+## unreleased
 - Made Lisp code tags more useful: the HTTP method for a request spec can now be determined by the result of a code tag.
 - Renamed customizable variable `verb-show-headers-buffer` to `verb-auto-show-headers-buffer`.
 - Added `verb-last` variable.