+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Renamed customizable variable `verb-show-headers-buffer` to `verb-auto-show-headers-buffer`.
 - Added `verb-last` variable.
 - Added `verb-headers-to-string` function.
+- Added `verb-post-response-hook` hook.
 
 ## **1.0.0** - 2020-01-13
 - Initial release.